[FAB-4821] Improve configtxlator reconfig examples
This improves the configtxlator README.md to use '.pb' as the suffix for
files contain protobuf messages rather than '.proto' which confuses the
'make protos' target.

It also improves the scripts for the existing examples, to use common
functions and adds an interactive mode which pauses between each step,
so that the user may inspect the file artifacts as they are generated.

It additionally improves error checking on the output of the assorted
commands, as the script would not reliably fail at the error points, but
might continue several statements.

Finally, it adds a new example for adding an organization to a channel.

Change-Id: I5c60cefe364f41dfc0d15c1dc5dd8f9ddc85ea2b
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>