[FAB-4848] Only allow TLS 1.2
Prior to this change, it was possible
to use TLS version < 1.2 to connect to
gRPC endpoints.  This change sets
the min/max TLS version to 1.2 for
the common gRPC server used by all
production gRPC endpoints.

It does not change the min version for
the various generic gRPC servers used
in other tests.

Change-Id: Ibf34777976551d12861599fb5ef37a93b07ece95
Signed-off-by: Gari Singh <gari.r.singh@gmail.com>