[FAB-4962] CouchDB username/passwords for end to end
The example for the end to end Getting Started scenario
should allow CouchDB and the peer to be configured with
CouchDB username and password.

The proposed change to docker-compose-couch.yaml would allow
admin usernames and passwords to be added for couchdb and peer
docker images.

All parameters added to the script are left uninitialized.
This allows the script to run "as is" and the CouchDB containers
are created in "admin party" mode.  If the new variables are
populated, the CouchDB containers and peers are created with
usernames and passwords.

Formatting updates to ledger.rst - make code block from core.yaml
render properly and tag environment variables.

Change-Id: I39595a6cd15aed1eed0b10f5fad8a3e09b08b9e5
Signed-off-by: Chris Elder <chris.elder@us.ibm.com>
Signed-off-by: Nick Gaski <ngaski@us.ibm.com>