[FAB-4901]: Harden delivery service unit tests.
In commit 74cc3b, delivery service was extended with timeout assertion
to check that client switched over to backup ordering service, while it
was a bit too small so client didn't have enough time to switch to alive
orderer endpoint. This commit also takes care to fix the bug of failover
waiting loop (missed for statement). Moreover adding a few more
assertions to enable spot the problem it case of test failure.

Change-Id: Ie33f4e6c0ac3998cc790778784d8cb732a333803
Signed-off-by: Artem Barger <bartem@il.ibm.com>
Signed-off-by: Gennady Laventman <gennady@il.ibm.com>