[FAB-5113] Clarify chaincode GetState doc
The current documentation does not make it clear what the behavior of
GetState is when the key does not exist.  This CR simply adds another
line to the documentation indicating that (nil, nil) will be returned so
that the chaincode developer does not have to guess or test this
behavior.

Change-Id: Ic610c3c488bd2b45482371231744c7a5579c7137
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>