[FAB-5150] remove bin data from install error
Authorization error contains bin data (package) which is useless
for error reporting and would get in the way of SDKs and other
tools's reporting. This CR removes the bin data. Further we don't
want to peek into the data on auth. failure and so will _not_
attempt to read the package to enhance the reported error.

. patch 2 fixed lscc unit test to look for right error string

Change-Id: Ic55eba8d890aa2489488fdbd54d4ccc3a8483fed
Signed-off-by: Srinivasan Muralidharan <muralisr@us.ibm.com>