[FAB-5261] Rm committers from blockcutter path
As one of the end goals of the change series for FAB-5258 the goal is to
differentiate between configuration transactions and standard ones.
When a message is known not to be a configuration message, it may be
processed with some simplifications.

The blockcutter leverages the filter.Committer interface to obfuscate
whether a message is a config transaction or not.  Consequently, this
notion must be removed in order to support the differentiation going
forward.

Ultimately, the entire notion of the filter.Committer will be removed.

Change-Id: I1466e14ce820db91930b685f48fec63a5d14cb15
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>