[FAB-5263] Reorganize orderer dir
The orderer directory has been overdue for being organized for some time
now.

With the change series of FAB-5258 there is a significant amount of code
being rearranged already, so it makes sense to organize the directory at
the same time, as the imports etc. will already be changing.

The new structure is very simple:

fabric/orderer - Contains the startup/main code

fabric/orderer/consensus - Contains the individual consensus plugins for
the orderer - ie Kafka/Solo, and eventually SBFT

fabric/orderer/common - Contains the common code which feeds
transactions to the consensus plugin, and satisfies the orderer
atomicbroadcast API definition

Change-Id: I112044dd5ade212378dab6c176265b76eb68ca80
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>