[FAB-5317] Cleanup filters package cruft
The filters package has been changed significantly over the course of
development, and especially so with the recent removal of the committers
concept.  This has left a lot of cruft which actually makes the package
less usable and less testable.

This CR removes all that cruft, resulting in a significant decrease in
lines of code, while improving the quality of the tests.

Change-Id: I3dc131b5775bde3b198b6f33438a076041a38ed0
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>