[FAB-4364] [FAB-5352] Support IP SANs in cryptogen
No matter what is provided as SANS in crypto-config.yaml, current code
logic of cryptogen takes them as host names, resulting in failed
communication if fabric network is configured with IP addresses.

The change proposes to examine the content of SANS, setting them as IP
SANs in the generated certificates if they are IP addresses.

Change-Id: Ie9cbc341ab21ec5966fdabcd48d79a9d05d7b961
Signed-off-by: Tony Yang <tony@arxanfintech.com>