[FAB-5274] Make WriteBlock asynchronous
In general, each consensus plugin for the ordering service has a single
thread which is responsible for receiving ordered messages, and writing
them into blocks.

This is a good thing, because it makes the consensus implementations
simpler and less error prone.  However, for operations like WriteBlock
which require some computations to be performed, followed by IO puts a
serious synchronous block in the path of these consensus threads.  While
the block is being written, the thread cannot do other things like drain
the queue or build the next block to write.

This CR decouples the actual writing of the block from the consenting
thread.  It accomplishes this by acquiring a lock in WriteBlock, then
spawning a go routine to go do the IO work, which releases the lock once
done.

While the lock is held, the consenter thread may perform other work,
but, if it gets to the point where it must construct a new block, the
CreateNextBlock function also acquires the lock, ensuring that the
integrity of the hash chain is not broken.

In non-scientific local testing with Solo, this produced an
approximate 30% increase in overall orderer throughput.

Change-Id: I11e2f9a716d415726c48066abe88a11a6e2fd1e0
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>