[FAB-5420] CouchDB docker image config improvements
Currently, the local.ini is placed in the /etc/local.d directory.
This file should be in the /etc directory based on the documentation:
http://docs.couchdb.org/en/2.0.0/config/intro.html

Also, the couchdb username and password are currently placed in
/etc/local.d/docker.ini.  When couchdb is started, the password is
hashed and stored in the /etc/local.ini.   The password in the
docker.ini is left in clear text.

Writing the username and password to the /etc/local.ini will prevent
passwords being left in clear text.

Change-Id: I0a6aca7092c000a99fe83720ec9dc18ec3372819
Signed-off-by: Chris Elder <chris.elder@us.ibm.com>