[FAB-5441] Add multisig support to peer CLI
Without utilizing the SDKs, it is currently not possible to submit a
configtx with signatures from multiple parties.  This makes it very
difficult if not impossible to modify things like channel membership.

This CR adds a new command to the peer channel CLI: signconfigtx

This command operates like the update command, but instead of signing
and submitting it to ordering, it simply signs the transaction, then
writes it back to disk.  This write is done in place, replacing the
input file.

Change-Id: I59bd922ec6ecaca11d8e23ef604228e51c229b46
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>