[FAB-5443] Have configtxlator emit default fields
Currently, configtxlator outputs configs with default fields omitted.
This means for instance that if a value has not been changed, its
version at 0 is not printed.  Similarly, if there is no CRL defined, the
CRL field is not printed.

Based on multiple reports, this behavior seems to be more confusing than
helpful.  This CR changes the code to turn the emit defaults flag on.

Change-Id: I4b393f4eb0473fd3e7ee09b7db6ccc4ac7c1932c
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>