[FAB-5445] Consolidate filters into msgprocessor
The individual filters are currently in subpackages of the msgprocessor
package.  However, the filters are simple enough this is probably no
longer warranted.

Additionally, in order to return more sane status codes, it's necessary
to have shared error definitions, which is difficult to do without
import cycles with the sub-package structure.

This CR moves the individual filter code to msgprocessor.

Change-Id: I6617f9b053345446d4636e1b62b7daf2280dee42
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>