[FAB-5236] System channel ID should be configurable
System channel ID is hardcoded to 'testchainid' when using provisional
genesis method. This should be configurable so that we could run multiple
tests against a stateful service, e.g. Kafka, without cleaning up the
environment after each run. Otherwise, after the first run, later ones
would recognize 'testchainid' and recover states from it. Also, we meant
to deprecate this hardcoded manner anyway.

Note: this is for testing purpose only, so we don't expose it in the
orderer config yaml.

Change-Id: I60590f52c304e003e3dd74c27aa0f3bec3bb996a
Signed-off-by: Jay Guo <guojiannan1101@gmail.com>