[FAB-5484] Fixes documentation for Windows users
With this change the documentation instructs Windows users
to work within the Docker Quickstart Terminal rather than
Git Bash which does not work for Windows 10.
This change also adds a note to the first curl command to
address the recurrent problem with Windows users reporting
failure due to using an outdated version of curl.

Patch-set #2: added a piece about installing npm grpc based
on users reporting this was necessary.

Change-Id: I95e7caeddc43e48f107955018aced112893ca540
Signed-off-by: Arnaud J Le Hors <lehors@us.ibm.com>