[FAB-5545] Allow building images from diff project
This change will allow fabric images and executables
to be built from a separate repository when fabric is
included as a git submodule.

By default, the build will assume that the repository is
hyperledger/fabric. When using fabric as a submodule, the
parent repository will need to set the PROJECT_NAME
environment variable.

Change-Id: Ia7a49052e1ebb2503631e047bb3b0957ea6ab704
Signed-off-by: Latitia M Haskins <latitia.haskins@gmail.com>