[FAB-5560] MSP manager can initialize with no MSPS
For no reason other than sanity checking, the MSP manager code currently
does not allow the MSP manager to be instantiated without any MSP
definitions.

Although this sounds sane, it actually causes a multitude of problems
any time the system is being run in a test or development mode,
attempting to exercise flows which do not depend on signature
verification.  This introduces an unnecessary special case (to handle an
uninitiatalized MSP manager) into the code, and in the worst case
results in nil pointer dereferences as as assorted components try to use
it.

This CR simply removes the checks which guards against the completion of
instantiation for an empty MSP manager.

This was originaly included as part of a CR series to change the
interface to the policy manager, but it's also applicable in this
series.

Change-Id: I524e796480744a00c80bdfbf366426cfdc5e1ab3
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>