[FAB-4370] Basic EndorserTx support in protolator
There have been numerous requests on rocketchat with people attempting
to inspect the endorser transaction contents.  Thanks to the complicated
structure of the endorser tx protos, it is very difficult for a casual
user to turn the binary blob into a human readable form.

This CR adds basic support for decoding typical endorser transactions.

Note: It makes no effort to handle transactions which have the proposal
visibility set to anything other than full, and no effort to support
endorser transactions which are not chaincode related (to my knowledge,
these are unimplemented).

Change-Id: I246b5a4ec60ae5748a4eedc064da54f1d3e92672
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>