[FAB-4974] Ledger proto messages enhancement for sidedb
This CR enhances the proto message for supporting the
hashed read-write set and private read-write set for the
private data. To summarize - under a namespace (chaincodeid),
there could be one or more sub-spaces (referred to as collections)
for managing private state. The hashed read-write set is intented
be part of the block and the private read-write set is intented to be
transferred to the need-to-have peers only.

Change-Id: I2f5127f93c945ab6559ed625f677d467e67bbcda
Signed-off-by: manish <manish.sethi@gmail.com>