[FAB-5646] Move channel config to config/channel
This is the first CR in a series designed to split the conflation of
configtx processing and channel configuration.  The bulk of the code in
fabric/common/config is specific to this channel configuration, so, this
CR moves the entirety of this directory to fabric/common/config/channel.

Ultimately, the common pieces will be extracted and moved back, but in
the interest of making things reviewable and minimizing the diff, it
makes sense to do this bulk move in one CR, then the move back in
another.

Change-Id: I124d48b0cbc7fd482e5118177ba9d59a81241d27
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>