[FAB-5647] Extract common comp from config/channel
In the last CR in these series, the entirety of the config structures
were moved to fabric/common/channel/config.  This includes code like the
Proposer and StandardValues which is applicable outside of the context
of the channel config.

This CR moves this subset back to the fabric/common/config directory.

Change-Id: I5ff5378a1362ee5da28deec2d46f5dd07ba889e0
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>