[FAB-5648] Extract Initializer from configtx
The configtx code depends upon an initializer, which provides resources
to the configtx code and is embedded to supply those resources to other
components.

At the time it was written, there was only one non-mock initializer in
the system, so it was stored within the configtx package.  However,
preparing to support other configtx processing, this code which is
dependent on the channel config must be removed from the configtx
package.

This CR simply moves the initializer code from configtx to
config/channel.

Change-Id: I03c91e68f39285bcd461eefec86d8943026de397
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>