[FAB-5535]: Add coordinator for blocks and pvtdata
Currently there is a state transfer layer in gossip, which takes care to
reordering the blocks according to the sequence number order, as well it
takes care to replicate blocks in case of network distruptions and there
are missing blocks in peer state.

In context of work required for FAB-1151, where there is private data
which distributed off the leadger, meaning not a part of transaction
proposal submitted to the ordering service. There is a need to
coordinate availability of this private data and the block itself.

Therefore this commit adds new entity to orchestrate the arrival of blocks
and the readiness of the block for commit, e.g. when peer will have all
relevant private data kept in transient store coordinator will commit
it into the ledger providing required private data.

Change-Id: I9cf47c9647e3f7fc949c5af2c6cd73c991f970d5
Signed-off-by: Artem Barger <bartem@il.ibm.com>