[FAB-5649] Invert configtx resource encapsulation
The configtx manager current retains a reference to the channel config,
but to support other config types, this actually needs to be inverted so
that the channel config retains a reference to the configtx manager.

This CR performs this inversion and fixes the embedding packages of the
configtx.Manager to instead embed the channelconfig.Resources.

[FAB-5650] Move channel creation from configtx

One of the last big offenders for putting channel config specific code
into the configtx package is the templating code, which provides
mechanisms to create new channel configurations.

The generic templating is fine to remain there, but the channel
specifics must be moved ot the channel config folder.

Note: These two sub-tasks were squashed into one because of otherwise
difficult to break import cycles.

Change-Id: Ib8bf829216e19fbbbd8841c9a3551356ed1676b6
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>