[FAB-5709] Infer `LastOffsetPersisted` correctly
The `LastOffsetPersisted` of block produced by Kafka orderer
may be off by 1 in a corner case, see FAB-5709 for details.

This is fixed by having `Ordered` method returning a boolean
indicating if there's still message pending in the receiver.

The old `ok` return value is dropped as `Ordered` does not
validate ingress message anymore and this value is always
true.

Also, a hacky way of calculating offset in subtests is removed
in favor of querying sarama mock Consumer directly.

Change-Id: Iaeac33cd60c27164e8bd732b6a61409f9b444f30
Signed-off-by: Jay Guo <guojiannan1101@gmail.com>