[FAB-5709] Infer `LastOffsetPersisted` correctly
The `LastOffsetPersisted` of block produced by Kafka orderer
may be off by 1 in a corner case, see FAB-5709 for details.

This is fixed by having `Ordered` method returning a boolean
indicating if there's still message pending in the receiver.

Also, a hacky way of calculating offset in subtests is removed
in favor of querying sarama mock Consumer directly.

Change-Id: I7f0384cd9df498b686d0971696fdcca102c52a59
Signed-off-by: Jay Guo <guojiannan1101@gmail.com>