[FAB-3949] refactor gossip UT waitgroup counter
In the gossip tests there is a waitgroup counter that is
used for the last test (that checks that goroutines ended)
to be able to run only when all the other tests have finished.

Currently the waitgroup counter is initialized to the number of
tests but it should be initialized in a more dynamic way that
shows which tests are disabled and which tests are not.

Change-Id: I491d7d3676aeb79fbe2a30fadc63b27a37f003c9
Signed-off-by: yacovm <yacovm@il.ibm.com>