[FAB-5806] Create channel config policy bundle
The channel configuration is currently processed via the configtx
package.  This is actually a conflation of responsibilities, as the
configtx package should be concerned about permission to modify a
configuration, not the actual contents of the change.

This CR creates a new channel config bundle, which will ultimately hold
an immutable copy of the channel configuration.  For the time being, it
implements a poor man's parsing of the policy tree.  This temporary
parsing will be removed in the future in favor of more robust parsing
once the mutability notion has been removed from the policy manager.

Change-Id: Ia3f4dc4d247afcc9c4e7895fff93cc2d007fae5a
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>