[FAB-5807] Add channel config values to bundle
The channel configuration values are currently mutable, which causes
problems.  Each sub-group of configuration data commits seprately,
leading to potentially inconsistent views of the channel configuration
when considered in the context of cross go-routine out of execution for
memory assignment.

This CR adds the channel config values to the immutable config bundle
and introduces a poor man's config value parser.  This parser will be
removed once the config values are entirely immutable, but is convenient
for now to remove an unwanted dependency on the configtx code.

Change-Id: I0295bd900677622bd9ab280386c1e8659e1f8d53
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>