[FAB-5808] Add configtx.Manager to immutable bndl
The immutable bundle does not need the configtx manager to actually
parse the config or policies, but it does need a way authorize new
configurations.  Therefore, the channel config needs to be fed to the
configtx.Manager, but without injecting a new dependency onto the
mutable logic.

This CR introduces a noop value and policy proposer so that the
configtx.Manager code can continue to authorize configuration updates
without injecting this new mutability dependency.  Once there are no
more dependencies on the configtx value/policy proposals, this noop
implementation will be deleted.

Change-Id: I4880602d9865fa6d38d136ea9562b4943ae2a41f
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>