[FAB-5814] Make policies.Manager immutable
In this change series to make the channel config immutable, up until
now, the focus has been on removing dependencies on the conflation
between the configtx authorization and the config parsing.  The real
payoff however is in removing the complicated state tracking logic from
all of the assorted mutable components.

This CR tackles the policies.ManagerImpl, which traditionally must
maintain a map of pending changes, create multiple implementations of
the manager, then stitch them together into a working hierarchy on
commit.  This is a great example of the simpliciation which comes from
not treating these structures mutably, as the complexity is lowered
substantially.

This also means that the temporary "poor man's policy parsing" may be
removed.

Change-Id: Ia5d0f507730fe5a230ddff1b7e987cdced460790
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>