[FAB-5819] Combine old/new channelconfig
This is a cleanup CR which consolidates the mixed references to the old
channel config code and the new channel config code.  In general, either
reference is simply changed to refer to fabric/common/channelconfig but
there are some package names which need cleanup as well.

Although large and scary looking, this is purely rearrangements of
imports and package names, no actual code logic is changed in this one.

Change-Id: Iddb2878782307ede3b2786f035673effe1425b86
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>