[FAB-5800] Allow orderer to set LogFormat backport
The orderer's logs always print using the default format string which
includes colors in the output.  Although helpful when looking at a
console, this makes logs difficult to handle programatically.

This CR adds a LogFormat option to orderer.yaml and uses it to
initialize the logging format for the orderer.

Change-Id: I7c595aa42e34255114953fab3dcbeafffc5cd377
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>