[FAB-5845] Separate trusted roots upon join channel
The current code assumes that the application orgs and the orderer orgs
are separate, but a peer and an orderer may be in the same org and then
this code might not (for example) populate the orderer TLS root CA certs.

The implication, is that a peer might not be able to connect to an orderer.

Change-Id: Ia19a45b42837d223282ce15402ebd5b4a3cb121b
Signed-off-by: yacovm <yacovm@il.ibm.com>
(cherry picked from commit cf82b4e70cc92fed3cf56374970a21cb60098541)
Signed-off-by: Gari Singh <gari.r.singh@gmail.com>