[FAB-5867] Make statedb validator to use bulkread
This CR makes the statedb validator to use bulk
read optimization APIs such as LoadCommittedVersions() and
GetVersion() to reduces the number of REST API calls between
peer and CouchDB to improve the performance.

Statedb validator first retrieves read set (including hashed reads) of all
transactions in the block (which is to be validated), then preloads the
committed version number of each key in the read sets into a cache (using
LoadCommittedVersion()) from the statedb, and finally compares
version of each key in read set against committed version (which is
retrieved from cache using GetVersion()) while validating a transaction.

Existing unit-test for validator is adequate to test this CR.

Change-Id: I96cc7fec2d7fcd07c7cffc5cb0aa07635e036164
Signed-off-by: senthil <cendhu@gmail.com>