[FAB-5868] Specify collection interface
This generic collection interface is implemented by different kinds of
collection types.

The collection's access policy interface is consumed by gossip
to govern access to private read-write sets.

The collection store acts as a collection factory and serves collection
objects based on collection criteria.

The nop-collection is an implementation of both interfaces, i.e. a
collection type, that allows all organizations to pull a private
read-write set and a collection store that just returns an instance of
this collection type.

Change-Id: I9a589c1609a719a593918896623586583568a262
Signed-off-by: Matthias Neugschwandtner <eug@zurich.ibm.com>