[FAB-830] - chaincode-level encryption
This change set introduces support for chaincode-level encryption. There are
two main interfaces: "EncShim" is similar to the chaincode shim in that it
exposes GetState and PutState functions to interact with the ledger.
Internally, encshim performs transparent calls to BCCSP as appropriately to
encrypt and decrypt data. Which keys and modes to use is governed by another
interface, "Entity" (and related). It captures a cryptographic entity that
can perform operations given one or more keys and modes. The chaincode
developer simply creates an EncShim instance (from the existing shim), an
Entity instance (equipped with appropriate crypto material and config) and
then can simply interact with the ledger by doing calls like

eshim.With(entity).PutState(key, val)

or

eshim.With(entity).GetState(key)

There are many ways in which a key can be passed to a chaincode (so that the
chaincode can create the Entity). One of them is using the Transient field.

Change-Id: I032dc711dd602e219be6ae924372e12dee93d3cd
Signed-off-by: Alessandro Sorniotti <ale.linux@sopit.net>