[FAB-5944] Restore disabled config update checks
In order to refactor the channel config to be immutable, checks which
depended upon the previous state had to be removed (in particular,
checks for the consensus type changing and for the org's MSPID changing
were temporarily removed).

This CR adds those checks back in a step which allows the previous
config to validate the new one via 'ValidateNew'.

Change-Id: Ic6c51f4b09f5aaaa8d176d850728287a1381b5e4
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>