[FAB-5973] CouchDB error when document deleted twice
CouchDB will throw an error if a document is deleted in more
than one update batch.  The first batch actually deletes the
document, but if the delete is included in a subsequent batch
an exception will be thrown since the document revision will
not be resolved correctly.   In this case, the batch update
should ignore documents that are already deleted.

Update documents are not impacted since the update operation
can correctly resolve the CouchDB revision info.

Change-Id: Ie4a329cd8d0be8578890657cdbabdb508f7cd3d6
Signed-off-by: Chris Elder <chris.elder@us.ibm.com>