[FAB-6069] Add capabilities package to common
There are capabilities which are peer specific, orderer specific, and
capabilities which are required of both (think, MSP types).  Each of
these capabilities need to be able to be checked.  This CR adds a
capabilities package to 'common' and skeleton code for checking the
/Channel group set of compatibility config.

For now, no capabilities are supported.

Change-Id: I5f3c9494d59abd82157284d7c8b255dcab08971b
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>