[FAB-6139] Add new channel config encoder package
configtxgen currently depends on the 'provisional' package for encoding
the information contained in configtx.yaml into the fabric native proto
format.  The provisional package was originally developed as part of the
orderer testing framework, and was never designed to work as it does
today.

This CR introduces a new encoder which more directly translates between
the viper config struct and the native proto format.  It relies on much
less indirection and should be easier for users to understand, easier to
hack on, and already has significantly better test coverage.

In addition, the new encoder allows for encoding of a subset of the
configuration, which is very useful for tests and tools.  There are many
places in fabric codebase which use the provisional package to generate
a genesis block, then unpack the the block, unpack the envelope, unpack
the config envelope, and finally access the config inside, and the new
package allows producing the config directly without these intermediate
steps.

This CR simply introduces the new encoder and tests.  It utilizes the
new simple util functions from the channelconfig and policies packages.

In subsequent CRs, the old provisional package will be removed, as will
the assorted old 'Template<...>' methods it depended upon.

Change-Id: I827504af301631171fc5a5998637b31e99c8dc4b
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>