[FAB-6141] Add more flexible channel create helper
The original channel creation tx generation code assumes that no
modifications have been made to the organization definitions in the
ordering system channel.  For longer running systems this assumption may
not hold.

Consequently, a new function which accepts the ordering system channel's
config is required so that it can set the correct versions for the
organization's in the channel creation transaction.  This CR adds this
new function.

Change-Id: Iebeb247bbea5c688a14c6133f793373da7357c9b
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>